mata: Increase doze brightness.
In a previous commit, the brightness for doze was set to the minimum possible
value (before the screen actually turns off completely) which basically rendered
the ambient display kind of useless as nothing was legible at that brightness.
This commit increases it to the point where it's not annoyingly bright
or dim.

Change-Id: Ibbd93f057f3a2221e9efbca7ccd9e43cdddb2efc